requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
anyio=3.5.0=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
bertopic=0.12.0=pypi_0
biopython=1.79=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blis=0.7.6=pypi_0
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.07.19=h06a4308_0
catalogue=2.0.6=pypi_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pypi_0
conllu=4.4.1=pypi_0
cryptography=37.0.1=py39h9ce1e76_0
csaps=1.1.0=pypi_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pypi_0
cymem=2.0.6=pypi_0
cython=0.29.28=pypi_0
datasets=2.3.2=pypi_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.5.1=pypi_0
en-core-sci-lg=0.4.0=pypi_0
en-core-sci-md=0.4.0=pypi_0
en-core-sci-scibert=0.4.0=pypi_0
en-core-sci-sm=0.4.0=pypi_0
en-ner-bc5cdr-md=0.4.0=pypi_0
en-ner-bionlp13cg-md=0.4.0=pypi_0
en-ner-craft-md=0.4.0=pypi_0
en-ner-jnlpba-md=0.4.0=pypi_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.6.0=pypi_0
fonttools=4.29.1=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
gensim=4.2.0=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
hdbscan=0.8.28=pypi_0
huggingface-hub=0.10.0=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.19.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39h06a4308_0
ipython=8.4.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
jpeg=9e=h166bdaf_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py39h06a4308_0
jupyter_client=7.2.2=py39h06a4308_0
jupyter_core=4.10.0=py39h06a4308_0
jupyter_server=1.17.1=py39h06a4308_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py39h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kaleido=0.2.1=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7f98852_1001
langdetect=1.0.9=pypi_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libwebp-base=1.2.2=h7f98852_1
lime=0.2.0.1=pypi_0
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
mpmath=1.2.1=pypi_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
murmurhash=1.0.6=pypi_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.3.0=py39h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.6=he412f7d_0
networkx=2.8.3=pypi_0
nltk=3.7=pypi_0
nmslib=2.1.1=pypi_0
notebook=6.4.11=py39h06a4308_0
numba=0.55.1=pypi_0
numexpr=2.8.3=py39h807cd23_0
numpy=1.21.0=pypi_0
numpy-base=1.21.5=py39ha15fc14_3
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathy=0.6.1=pypi_0
patsy=0.5.2=py39h06a4308_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=pypi_0
pip=22.1.2=py39h06a4308_0
plotly=5.6.0=pypi_0
preshed=3.0.6=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=pypi_0
pybind11=2.6.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pynndescent=0.5.6=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py39heee7806_0
pysbd=0.3.4=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py39h06a4308_0
pywavelets=1.3.0=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=23.2.0=py39h6a678d5_0
readline=8.1.2=h7f8727e_1
regex=2022.1.18=pypi_0
requests=2.28.1=py39h06a4308_0
responses=0.18.0=pypi_0
sacremoses=0.0.47=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
scispacy=0.4.0=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart-open=5.2.1=pypi_0
sniffio=1.2.0=py39h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
spacy=3.0.8=pypi_0
spacy-alignments=0.8.4=pypi_0
spacy-legacy=3.0.9=pypi_0
spacy-transformers=1.0.4=pypi_0
sqlite=3.38.5=hc218d9a_0
srsly=2.4.2=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=py39h7f8727e_0
sympy=1.10=pypi_0
tenacity=8.0.1=pypi_0
terminado=0.13.1=py39h06a4308_0
testpath=0.6.0=py39h06a4308_0
thinc=8.0.13=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.5.4=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=pypi_0
torch=1.10.2=pypi_0
torchaudio=0.10.2=pypi_0
torchvision=0.11.3=pypi_0
tornado=6.1=py39h27cfd23_0
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.9.2=pypi_0
typer=0.3.2=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
umap-learn=0.5.2=pypi_0
urllib3=1.26.9=py39h06a4308_0
wasabi=0.9.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
xlsxwriter=3.0.3=pypi_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=pypi_0
zeromq=4.3.4=h2531618_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.0=ha95c52a_0
